Create generator file for algebra question
